It says "operations to shuffle the deck and to deal one card. While not a requirement today, a likely future enhancement is the need to deal all the cards in a deck."
Should only one card be dealt or should the whole deck be dealt when deal is called? Is the future enhancement likely to be to return all cards in a deck at once?

Basically what does "deal" mean? Give one card from the deck, always the first card (that is what I implemented) or give each card in the deck one at a time?